[Podspec] Remove circular dependency
You do not need to explicitly depend on a subspace since by default it will be included. The subspec will contain the dependency from it's parent spec and therefore lead to a circular dependency.

As noted in https://github.com/CocoaPods/Molinillo/issues/6, the coming version of CocoaPods will now allow this.